% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{hadza}
\alias{hadza}
\title{Hadza gut microbiome dataset}
\format{
A `phyloseq` object named \code{physeq}.
}
\source{
\url{https://www.nature.com/articles/ncomms3654}
}
\usage{
hadza
}
\description{
A sample phyloseq object from the Hadza gut microbiome, used as a demo.
}
\keyword{datasets}
